Update tests and executable list for bitboard
The source file bitboard.cpp has been added to the executable
list for bitboard_test in CMakeLists. The lack of this file in
the executable list was preventing the project from build. With
this addition, the project now builds successfully.

The single test in bitboard_test (NorthOne) has been updated to
follow the Assemble - Act - Assert model.