Add tests for bitboard operations
Tests have been added for north, east, south, and west bitboard
operations. For each of these operations two test cases are now
available:

- Using the operations to shift bits from a rank or file in the
middle of the board. This means the bits will still all be on the
board after the operation.
- Using the operations to shift bits from a rank or file at the
edge of the board. This means that none of the bits will be on the
board after the operation.