Add tests for all but one bitboard.h functions
Some functions have two test cases. The functions scan_forward and
scan_backward are tested with a single bit on the bitboard and also
with an empty bitboard.

The functions rank_index and file_index are tested both with a valid
square index (square 7), and with an invalid square index (65).

The only function in bitboard.h that is not tested is print_board.
This function has been left untested because it will soon be removed
in favor of a Model-View-Controller pattern for the UI.