Add tests for Rays with blocker on adjacent square
Adds a single test to each of the existing eight test suites in
attackmap_tests.cpp (i.e., adds eight tests in all). These tests ensure
that a ray is correctly returned when a blocking piece is placed on a
square immediately adjacent to the square of the ray casting piece. For
example, a piece that casts an east ray placed on G5 with a blocking
piece on F5.

The AttackMapTest test suite has been renamed NorthRayAttackMapTest to
be consistent with the other test suites in the file. The NorthRay
instance that is created in this test suite has also been changed to use
the auto type, also for consistency with the rest of the tests in the
file.