Update Move to implement a command pattern
The move class has been implemented to implement a command pattern.

All moves are serial moves from now on, these will be specified by
the source square index and the target square index. As a result of
ve to implement a command pattern

The move class has been implemented to implement a command pattern.

All moves are serial moves from now on, these will be specified by
the source square index and the target square index. As a result of
this the SerialMove class has been removed, and the Move class has
been updated to use the type Square for source and target (instead
of type Bitboard).

The Move will be performed using a command pattern `execute()` method.
As such there is no need for the `get_source()` and `get_target()`
methods and these have been removed. In addition, there is no need for
`target` to be protected, it has been made private.

In order to implement the command pattern, the Move class needs to
have some reference to the object on which it will execute its command.
This will be some class derived from the Pieces class. This has been
added to the arguments of the constructor, to the private attributes
list, and the initialiser list of the constructor. It is added as a
std::shared_ptr<Pieces>.

The `execute()` method has been added to the Move class. The current
implementation of this method requires a public method
`Pieces::make_move(Square source, Square target)`. This method has
been added to the Pieces class.

Using this method of implementing moves causes problems with the
existing Pieces implementations. All of these issues will need to
be dealt with. Sliding pieces have been completely commented out
from now, and the new Move class will be implemented only for pawns
for now.

The two pawn methods: `single_push()` and `double_push()` have been
modified to only return a source bitboard. They were returning a Move,
but this only makes sense for the old Move implementation. Now these
function only return a single bitboard of pawn source squares. Thus,
`single_push()` returns a bitboard of all pawns able to push forward
one square, while `double_push()` returns a bitboard of all pawns
able to push forward two squares. The previous implementation created
both the source and the target bitboard and then place them in an old
Move object.

West and east captures have been commented out.

A new class called SourceTargetPair has been created to replace Move
in the pieces.cpp and pieces.h files. In trying to implement the new
Move class I have not found a way to make it possible to create the
Move commands from within the class itself. Thus, the plan for now is
to have Piece classes return stacks of SourceTargetSquares. Then,
whatever client calls the piece to get that stack of squares can
organise it into Move command objects itself.

The files main.cpp and pieces_test.cpp have been modified to account
for the new changes that have been made. These changes are either
changing types to match new types or just commenting out code that
needs to be updated.
this the SerialMove class has been removed, and the Move class has
been updated to use the type Square for source and target (instead
of type Bitboard).

The Move will be performed using a command pattern `execute()` method.
As such there is no need for the `get_source()` and `get_target()`
methods and these have been removed. In addition, there is no need for
`target` to be protected, it has been made private.

In order to implement the command pattern, the Move class needs to
have some reference to the object on which it will execute its command.
This will be some class derived from the Pieces class. This has been
added to the arguments of the constructor, to the private attributes
list, and the initialiser list of the constructor. It is added as a
std::shared_ptr<Pieces>.

The `execute()` method has been added to the Move class. The current
implementation of this method requires a public method
`Pieces::make_move(Square source, Square target)`. This method has
been added to the Pieces class.

Using this method of implementing moves causes problems with the
existing Pieces implementations. All of these issues will need to
be dealt with. Sliding pieces have been completely commented out
from now, and the new Move class will be implemented only for pawns
for now.

The two pawn methods: `single_push()` and `double_push()` have been
modified to only return a source bitboard. They were returning a Move,
but this only makes sense for the old Move implementation. Now these
function only return a single bitboard of pawn source squares. Thus,
`single_push()` returns a bitboard of all pawns able to push forward
one square, while `double_push()` returns a bitboard of all pawns
able to push forward two squares. The previous implementation created
both the source and the target bitboard and then place them in an old
Move object.

West and east captures have been commented out.

A new class called SourceTargetPair has been created to replace Move
in the pieces.cpp and pieces.h files. In trying to implement the new
Move class I have not found a way to make it possible to create the
Move commands from within the class itself. Thus, the plan for now is
to have Piece classes return stacks of SourceTargetSquares. Then,
whatever client calls the piece to get that stack of squares can
organise it into Move command objects itself.

The files main.cpp and pieces_test.cpp have been modified to account
for the new changes that have been made. These changes are either
changing types to match new types or just commenting out code that
needs to be updated.