Start major redesign of pieces
The current design is causing problems. One of this problems
is that pawns and other pieces need to be handled differently,
and, therefore it is not clear that they can have the same
interface.

The design will proceed assuming that Pawns will be uses a
different interface to other pieces.

The main.cpp file has been modified to get rid of everything
except pawns.

In pieces.cpp and pieces.h a Colour class has been created
that (it is hoped) will simplify dealing with pieces of other
colours. Each instance of the Pawn class will own a std:shared_ptr
to an a colour derived class (Black or White).