Refactor tests to match Pawns interface
Three of four test have been refactored. The one test that hasn't
been refactored is ReturnVaildCaptures for WhitePawns, this is
because captures have not been implemented in the Pawns class yet.

A small bug fix has been made in the Pawns class. The
`all_quiet_moves()` method was not returning the correct result for
double push targets and was instead returning the single push
target bitboard. This has been corrected by using the
`pawn_double_push_target()` method from the Colour class.