Add ChessMen class and make_move() method
ChessMen class has been added. This class holds all of the pieces
for one side of a chess game.

A Piece class has been created as an abstract interface for all
pieces. A single pure virtual method (make_move()) has been added
to this class. The Pawns class has been updated to inherit from the
new Piece class and the make_move() method has been implemented in
this class.

Due to the way in which Pawns and Colour are implemented. A
make_move method has also been added to the Colour class. The
Piece::make_move() method updates that pieces bitboard and the
empty_squares bitboard. The Colour::make_move() method updates
the bitboard for all pieces of a particular colour.