Implement way to get piece on a specified square
In order to implement captures, the Move object that encapulates the
move command must have some reference to the piece type that is being
captured. The Move object requires this reference so that, when the
move command is executed, it can remove the captured piece from the
bitboard of the piece type being captured.

The Pawns class can return all the squares that are currently capture
targets, i.e., a bitboard of all the squares that are a) attacked by a
pawn, and b) contain an enemy piece to capture. This bitboard can be
serialised to obtain a list of all squares with enemy pieces that can be
captured.

A method has been added to the ChessMen class that returns the piece
that is currently on a given square, by providing the index number of
that square. Thus, the move generator can take all of the square
indicies of attacked square and get the reference to the piece type
object by using ChessMen::get_piece(square_index).

A NullPiece class has also been added. If the square is empty when
get_piece is called then the method will return a NullPiece.