Implement pseudo legal move generation for pawns
Generation of pseudo legal moves in an important step in generating the
list of legal moves that are available on a single move of chess game.

The MoveGenerator has been given methods for generating pseudo legal
moves for pawns. These methods consist of a public method for clients to
extract pseudo legal moves from the generator and a number of private
methods for supporting the public `generate_pseudo_legal_moves()`
method. Moves are returned as a std::stack<std::shared_ptr<Move>>.

An additonal Move derived class has been added: `PseudoLegalCapture()`
that implements the Move interface for capture moves.

The interface to the Piece class has been extended it include methods
for removing captured pieces from the Piece bitboard and checking if a
piece is the NullPiece. The extended interface has been extended in both
Piece derived classes (NullPiece and Pawns).