Remove pointer to Pawns bitboard
The bitboard used by the Pawns class needed to be created outside the
class (as a std::shared_ptr<Bitboard> and given to it through the
constuctor. This meant there there was potential for this pointer to
remain owned by a user outside the Pawns class, and potential for its
state to be changed by that user without access to the Pawns object at
all.

The constructor Pawns class has been modified and a second constuctor
added. The original constructor no longer takes the board by reference
but by value, while the new constructor allows for the Pawns object to
be created without providing a bitboard at all. In the new constructor,
if the object is created with out specifing a bitboard of piece
positions then the board is initialised as empty (Bitboard(0)).

Two methods have been added to the class, one that returns the current
state of the bitboard (this is a method that was added to Pawns base
class Piece, and is overridden by Pawns). The second added method allows
for bitboard to be set to the standard starting position of the Pawns
(either rank 2 for white pawns, or rank 4 for black pawns). Due to the
different starting positions of white and black pieces this is
implemented by adding methods to the classes derived from the Colour
class.

Tests have been updated to reflect the changes to the interface of the
Pawns class.