Add set_square function for Bitboard
This commit adds a helpful `set_square` function to the `bitboards` namespace.
The function has two prototypes:

 - void set_square(Bitboard& board, Square square)
 - void set_square(Bitboard& board, Rank rank, File file)

These functions allow for a single square in the bitboard to be set to true,
either by reference to a square index or by reference to a rank index and a file
index. The bitboard is passed by reference and the modification will be applied
to the bitboard as a side effect.

If the square is already set to true the Bitboard will remain unchanged.