Fix error in pawn attack source generation.
Modifications were made due to a failing test:
`BoardMoveGenerationTests`: `CanGenerateAndMakeEpCapture`.

Created eight new tests. These tests are for pawn attack sources and
targets in both east and west directions for both black and white pawns.
Four out of the eight tests failed:

-	WhitePawnsTest_EastAttackSources
-	WhitePawnsTest_WestAttackSources
-	BlackPawnsTest_EastAttackSources
-	BlackPawnsTest_WestAttackSources

The problem was in the functions for generating east and west attack
sources, this has been corrected.

In addition, documentation has been added to the pawn source/target
generation functions in pieces.h.

The changes in this commit allow the `CanGenerateAndMakeEpCapture` test
to pass.