Fix processing of en passant captures.
This change was made due to failing test: `MoveTest`: `CanMakeEpCapture`.
After running the debugger, it was found that the `process_move`
function would call the `process_ep` function correctly if the move is
an ep capture. However, it then continued to execute the rest of the
`process-move` function after `process_ep` returned. This is not the
desired behaviour.

To fix this issue `process_capture` now returns after the call to
`process_ep` returns.